requirements.txt
openrouter>=0.8,<1.0
pandas>=2.0,<4.0
python-dotenv>=1.0,<2.0
sqlparse>=0.5,<1.0
